+++ src/main/kotlin/org/example/coha/domain/post/service/FileStorageService.kt
@@ -8,8 +8,7 @@ interface FileStorageService {
 
     // MultipartFile을 받아 파일을 저장하고, 저장된 파일의 경로를 반환하는 메소드
     fun storeFile(file: MultipartFile): String
+    fun storesupaFile(file: MultipartFile): String
     // 주어진 파일 이름을 기반으로 파일을 로드하고 Resource 형태로 반환하는 메소드
     fun loadFile(filename: String): Resource
-
-    fun storesupaFile(file: MultipartFile): String
 }
